# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skillset API",
    "version": "0.1.0",
    "description": "A self-hosted Registry for Agent Skills. This is the contract the CLI depends on; the web interface uses the same routes and has no private data path of its own.\n\n**Conventions.** Every JSON key on the wire is snake_case, which means wire keys line up with Postgres column names. TypeScript types, interfaces, and classes are camelCase, so conversion happens once at the boundary in the shared module rather than ad hoc per call site.\n\n**Artifact bytes never pass through this API.** Publishing returns a presigned upload target that the caller writes to directly, and reading an Artifact redirects to a short-lived presigned location. The API therefore validates the metadata a caller claims about a Skill, never the Artifact itself (ADR-0001)."
  },
  "servers": [{ "url": "/api" }],
  "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
...
  "paths": {
...
    "/identity-providers": {
      "get": {
        "tags": ["identity-providers"],
        "summary": "List every configured Identity Provider",
        "description": "Requires at least the admin role. Includes disabled Providers, and never includes a client secret.",
        "responses": {
          "200": {
            "description": "Every configured Identity Provider",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items"],
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/IdentityProvider" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/AdminOnly" }
        }
      },
      "post": {
        "tags": ["identity-providers"],
        "summary": "Configure an Identity Provider",
        "description": "Requires at least the admin role. At most one Provider per kind (ADR-0017). A Provider may be created enabled, but only with a permitted organisation set — with just-in-time provisioning that gate is the only control on who gets an account (ADR-0015).\n\nThere is no issuer URL to supply: Better Auth knows each kind's endpoints (ADR-0016), so what an Admin configures is the credential pair and the organisation to admit.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["kind", "display_name", "client_id", "client_secret"],
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": ["google", "microsoft", "github"],
                    "description": "The Provider's identity, and immutable after creation (ADR-0017)."
                  },
                  "display_name": { "type": "string", "minLength": 1 },
                  "client_id": { "type": "string", "minLength": 1 },
                  "client_secret": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Stored as given, and never returned by any response."
                  },
                  "permitted_organisations": { "type": "array", "items": { "type": "string", "minLength": 1 }, "maxItems": 50, "description": "Organisations this Provider admits — Workspace domains, Entra tenant ids, or GitHub organisation logins by kind. A login is admitted if it matches any one of them, case-insensitively. An empty array means no organisation check runs at all and everyone the provider authenticates gets a reader account (ADR-0021)." },
                  "enabled": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The configured Identity Provider",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IdentityProvider" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/AdminOnly" },
          "409": {
            "description": "An Identity Provider of that kind is already configured (`kind_taken`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/identity-providers/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/IdentityProviderId" }],
      "patch": {
        "tags": ["identity-providers"],
        "summary": "Change an Identity Provider's configuration",
        "description": "Requires at least the admin role, and is also how a Provider is enabled and disabled. `kind` cannot be changed: it is the Provider's identity (ADR-0017), so altering it would silently repoint a live configuration rather than configure a second one. Omitting `client_secret` leaves the stored one untouched, which is what lets a Provider be edited without its secret ever being read back.\n\nChanges reach logins by two routes, and the difference matters (ADR-0019): `enabled` and `permitted_organisations` are read on every login and take effect immediately, while new credentials take effect once the auth instance is rebuilt on the next request to its routes. A stale instance can therefore fail a login through a Provider added moments ago, but can never admit someone a disabled Provider should have refused.\n\nThere is deliberately no delete route. Disabling is how a Provider is taken out of service — deleting one would lock out everyone who signs in through it over a mistyped click.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "display_name": { "type": "string", "minLength": 1 },
                  "client_id": { "type": "string", "minLength": 1 },
                  "client_secret": { "type": "string", "minLength": 1 },
                  "permitted_organisations": { "type": "array", "items": { "type": "string", "minLength": 1 }, "maxItems": 50, "description": "Organisations this Provider admits — Workspace domains, Entra tenant ids, or GitHub organisation logins by kind. A login is admitted if it matches any one of them, case-insensitively. An empty array means no organisation check runs at all and everyone the provider authenticates gets a reader account (ADR-0021)." },
                  "enabled": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated Identity Provider",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IdentityProvider" } } }
          },
          "400": {
            "description": "Invalid request body",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/AdminOnly" },
          "404": { "$ref": "#/components/responses/IdentityProviderNotFound" }
        }
      }
    },
...
    "/connections": {
      "get": {
        "tags": ["connections"],
        "summary": "List your own Connections",
        "description": "Requires at least the `writer` role. Importing exists to publish, and a reader cannot publish — so a reader is refused rather than shown an empty list, and no connect action is offered to them.\n\nReturns only the caller's own Connections. No token appears in this shape, and there is no Admin variant that returns one: the tokens exist solely for the Registry to spend on that writer's behalf.",
        "responses": {
          "200": {
            "description": "The caller's own Connections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items"],
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/Connection" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/WriterOnly" }
        }
      }
    },
    "/connections/{provider}/start": {
      "parameters": [
        { "$ref": "#/components/parameters/GitProvider" },
        {
          "name": "integration_id",
          "in": "query",
          "required": false,
          "schema": { "type": "string", "format": "uuid" },
          "description": "Which Integration (app) to connect through, when the provider has more than one configured (ADR-0025). Omit when it has exactly one — there is nothing to choose."
        }
      ],
      "get": {
        "tags": ["connections"],
        "summary": "Begin connecting a Git Provider",
        "description": "Requires at least the `writer` role. Redirects to the provider's **installation** page rather than a plain authorize URL, which is what lets the writer choose which repositories the Registry may read in the same pass as authorizing it — and why the grant is an owner-chosen list rather than everything they can read.\n\nSets an httpOnly, `SameSite=Lax` cookie holding a nonce. `Lax` is required rather than chosen: the callback arrives as a top-level GET navigation from the provider, and `Strict` would withhold the cookie on exactly that request. The nonce is what makes the state single-use — a state lifted from a log or a referrer is useless in another browser.\n\nNot a JSON route: the response is a 302, and a browser is expected to follow it.",
        "responses": {
          "302": {
            "description": "Redirect to the provider's installation page, with a signed state",
            "headers": {
              "Location": { "schema": { "type": "string" } },
              "Set-Cookie": {
                "description": "The connect flow's nonce: httpOnly, SameSite=Lax, five-minute lifetime.",
                "schema": { "type": "string" }
              }
            }
          },
          "400": {
            "description": "That Git Provider has no credentialed flow at all, or more than one Integration is configured for it and `integration_id` was not given",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/WriterOnly" },
          "409": { "$ref": "#/components/responses/IntegrationNotConfigured" }
        }
      }
    },
    "/connections/{provider}/callback": {
      "parameters": [{ "$ref": "#/components/parameters/GitProvider" }],
      "get": {
        "tags": ["connections"],
        "summary": "Complete a Connection from the provider's callback",
        "description": "Requires at least the `writer` role. The provider redirects the writer here with `code` and `state`.\n\nThe state is verified and the nonce matched **before anything else** — no database read, no outbound request. This is the CSRF surface of the flow: without it an attacker could walk a victim's browser through their own authorization and attach their GitHub account to the victim's Registry account. A state that is missing, tampered with, expired, for another provider, minted for another User, or whose nonce does not match the browser's all yield the same `invalid_state`, because naming the failed check would tell an attacker which part of their forgery to fix.\n\nThe nonce cookie is cleared on every path out of this route, successful or not. That is what burns the state.\n\nOn success the grant is upserted on `(user_id, provider)` — reconnecting replaces rather than accumulating — and the writer is redirected back to settings.",
        "parameters": [
          { "name": "code", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "state", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "302": {
            "description": "Connection stored; redirect back to settings",
            "headers": { "Location": { "schema": { "type": "string" } } }
          },
          "400": {
            "description": "The state could not be verified, or the provider is not connectable",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/WriterOnly" },
          "409": { "$ref": "#/components/responses/IntegrationNotConfigured" },
          "502": {
            "description": "The provider refused to exchange the code, or would not identify the account",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/connections/{provider}": {
      "parameters": [{ "$ref": "#/components/parameters/GitProvider" }],
      "delete": {
        "tags": ["connections"],
        "summary": "Disconnect a Git Provider",
        "description": "Requires at least the `writer` role. Deletes the Connection and nothing else — in particular it does not touch the writer's sign-in account, so withdrawing repository access can never lock anyone out.\n\n**This is Registry-local.** It stops *this Registry* using the grant; it does not withdraw the grant at the provider, which the writer must do in their own provider settings. Interface copy must say so — anything else is untrue.",
        "responses": {
          "204": { "description": "Disconnected" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/WriterOnly" },
          "404": {
            "description": "The caller holds no Connection to that Git Provider",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Signed session cookie, http-only, secure, strictly same-site. Carries only the User's id; the role is resolved from the database on every request (ADR-0005). Cannot be revoked before expiry."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A Token, used by the CLI. Looked up by hash on every request; revocable immediately."
      }
    },
...
    "parameters": {
...
      "GitProvider": {
        "name": "provider",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "enum": ["github", "gitlab"] },
        "description": "A Git Provider this Registry reads from."
      },
      "IdentityProviderId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "The Identity Provider's id.",
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "responses": {
...
      "IdentityProviderNotFound": {
        "description": "No Identity Provider by that id",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthenticated": {
        "description": "No valid session or Token. The web interface treats this as an expired session, discards cached state, and returns to login.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "WriterOnly": {
      "description": "The caller's role is below writer. Importing exists to publish, and a reader cannot publish.",
      "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
    },
    "IntegrationNotConfigured": {
      "description": "No Integration is configured for that Git Provider, so there is no credential to connect with. An Admin's problem: an Integration row's existence is the only switch Importing has (ADR-0024).",
      "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
    },
    "AdminOnly": {
        "description": "Requires at least the admin role (admin or the superadmin)",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "ValidationFailed": {
        "description": "The request failed a validation rule. `code` names the rule and `field` names what broke it, so the CLI and the web interface can report the same specific reason.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
...
      "IdentityProvider": {
        "type": "object",
        "description": "An Admin's view of a configured Identity Provider. `client_secret` is deliberately absent from this and every other response: it is written and never read back, so no response can leak it, not even to the Admin who set it.",
        "required": [
          "id",
          "kind",
          "display_name",
          "client_id",
          "permitted_organisations",
          "enabled",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "kind": { "type": "string", "enum": ["google", "microsoft", "github"] },
          "display_name": { "type": "string" },
          "client_id": { "type": "string" },
          "permitted_organisations": {
            "type": "array",
            "items": { "type": "string" },
            "description": "The organisations allowed to sign in: Workspace domains (Google `hd`), tenant ids (Entra `tid`), or GitHub organisation logins. A login is admitted if what the Provider asserts matches any one of them, case-insensitively — a claim for the first two, a live `/user/orgs` check for GitHub, which issues no ID token (ADR-0018) — and never against the email address's suffix. An empty array is a configuration rather than a blank: no check runs, and everyone the provider authenticates gets a reader account here (ADR-0021)."
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the login page offers it and logins through it are accepted. Disabling is how a Provider is taken out of service; there is no delete."
          },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Connection": {
        "type": "object",
        "description": "A writer's own grant of repository access (ADR-0024). There is no token field, and no variant of this shape has one — a Connection's tokens never leave the server in any response, on any route.",
        "required": ["provider", "integration_id", "external_account_login", "created_at", "updated_at"],
        "properties": {
          "provider": { "type": "string", "example": "github" },
          "integration_id": {
            "type": "string",
            "format": "uuid",
            "description": "Which Integration (app) this grant was issued through — a provider may have more than one configured (ADR-0025)."
          },
          "external_account_login": {
            "type": "string",
            "example": "ada-work",
            "description": "The connected account at the provider. Present because a writer needs it: the connected account need not be the one they sign in with, and which account is connected is usually the answer to why an Import cannot see a repository."
          },
          "created_at": { "type": "string", "format": "date-time", "description": "When the grant was given — the consent's own timestamp." },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "description": "A stable machine-readable identifier for what went wrong — for validation failures, the specific rule that failed."
              },
              "message": { "type": "string" },
              "field": {
                "type": "string",
                "description": "The field or file that broke the rule, when there is one."
              }
            }
          }
        }
      }
    }
  }
}